+++ tools/deletebyquery/deletebyquery.py
@@ -89,6 +89,8 @@ def delete_by_query(args):
     else:
         raise RuntimeError("either query or jsonparams is required")
 
+    query.commonparams.rows(args.rows)
+
     if check_query(query):
         logging.info("Collecting tiles ....")
         solr_docs = do_solr_query(query)
@@ -259,6 +261,13 @@ def parse_args():
                         required=False,
                         choices=['1', '2', '3', '4', '5'],
                         default='3')
+
+    parser.add_argument('--solr-rows',
+                        help='Number of rows to fetch with each Solr query to build the list of tiles to delete',
+                        required=False,
+                        dest='rows',
+                        default=1000,
+                        type=int)
 
     return parser.parse_args()